LayoutTests/canvas/philip/tests/2d.strokeRect.globalcomposite-expected.txt

< [index] >
2d.strokeRect.globalcomposite
strokeRect is not affected by globalCompositeOperation
References: 2d.strokeRect
Actual output:
Expected output:

Passed
